Introduced SQLite functionality.
All stocks are now stored in the SQLite database. Routines include a way to initialize the database and import stocks from the stock_list.json file. Also made updates to the SVR routines to fit in with the database requirements and to give out more helpful output for logging. Need to implement a check for extremely high/low prices as this seems to be where the SVR model tends to struggle.