port_trader and bug fixes
Introduced a portfolio check that will check each position on each run. If the position has not been traded within the time limit and has a net gain or too heavy of a net loss, the position is closed. Implemented to try to limit more severe losses and collect gains more quickly. 
Additionally: 
- Fixed bugs in the min/max checker and SVR trader routines. 
- Added a method to get individual stock position information from Alpaca.